requirements.txt
pynetbox==7.3.3
pyvmomi==8.0.3.0.1
requests==2.32.3
deepdiff==8.0.1
pyyaml==6.0.2
redis
ldap3
netboxlabs-diode-sdk
python-hcl2>=4.3.0